Phase 2: stream layer reorganization (no behavior change)
Split the overloaded internal/io_helpers.py into a cohesive
internal/streams/ package and relocate the decompressor stream modules
into it, with no public-API or behavior change (test suite is unchanged):

- New internal/streams/ package:
  - compat.py  — type/seek shims, read_exact, is_seekable, BinaryIOWrapper,
    is_filename/is_stream, ensure_binaryio/ensure_bufferedio, open_if_file
  - detect.py  — RecordableStream, RewindableStreamWrapper
  - slice.py   — SlicingStream
  - stats.py   — IOStats, StatsIO
  - concat.py  — ConcatenationStream
  - errors.py  — ExceptionTranslatorFn, ErrorIOStream, run_with_exception_translation
- Relocate decompressor modules (git mv, history preserved):
  formats/decompressor_stream.py -> streams/decompress.py,
  formats/xz_stream.py -> streams/xz.py,
  formats/lzip_stream.py -> streams/lzip.py; archive_stream.py stays put.
- Simplify BinaryIOWrapper: drop the per-call method-replacement hot-path
  trick (self.read = self._raw.read, ...) in favor of plain delegation,
  preserving the same fallbacks and exceptions.
- Reduce io_helpers.py to a 45-line re-export shim so backends and tests
  importing from archivey.internal.io_helpers stay untouched.

Verified: pytest identical to Phase 1 baseline (4172 passed, 3013 skipped,
43 xfailed), mypy --strict clean (36 files), ruff clean.
