+++ src/archivey/internal/streams/__init__.py
@@ -1,0 +1,1 @@
+"""Internal stream utilities: I/O compatibility shims, wrappers, and decompressors."""
