+++ src/archivey/internal/streams/xz.py
@@ -27,7 +27,7 @@ from dataclasses import dataclass
 from typing import BinaryIO, cast
 
 from archivey.exceptions import ArchiveCorruptedError, ArchiveEOFError
-from archivey.formats.decompressor_stream import (
+from archivey.internal.streams.decompress import (
     SeekPoint,
     _SegmentedDecompressorStream,
 )
